move patient-summary page components to general components folder
