+++ src/app/components/PatientSummaryViewer/generics/DetailedTable.tsx
