+++ src/app/components/PatientSummaryViewer/generics/TabSection.tsx
