+++ src/app/components/PatientSummaryViewer/generics/resource.types.ts
