+++ src/app/components/PatientSummaryViewer/resources/AllergyIntolerance/index.tsx
