+++ src/app/components/PatientSummaryViewer/resources/Composition/index.tsx
