+++ src/app/components/PatientSummaryViewer/resources/Condition/index.tsx
