+++ src/app/components/PatientSummaryViewer/resources/MedicationStatement/index.tsx
