+++ src/app/components/PatientSummaryViewer/resources/Observation/index.tsx
