+++ src/app/components/PatientSummaryViewer/resources/Organization/index.tsx
