+++ src/app/components/PatientSummaryViewer/resources/Patient/index.tsx
