+++ src/app/components/PatientSummaryViewer/resources/Practitioner/index.tsx
